# feat(query): optimize StringSearchLike vector_const