--- a/src/query/functions/src/scalars/comparisons/comparison_like.rs
+++ b/src/query/functions/src/scalars/comparisons/comparison_like.rs
@@ -76,21 +76,24 @@ impl StringSearchImpl for StringSearchLike {
             PatternType::PatternStr => {
                 let pattern = simdutf8::basic::from_utf8(rhs)
                     .expect("Unable to convert the LIKE pattern to string: {}");
-                let mut sub_strings: Vec<&str> =
-                    pattern.split(|c: char| c.is_ascii_punctuation()).collect();
+                let mut sub_strings: Vec<&str> = pattern
+                    .split(|c: char| c == '%' || c == '_' || c == '\\')
+                    .collect();
                 sub_strings.retain(|&substring| !substring.is_empty());
+                let is_empty = sub_strings.is_empty();
                 let re_pattern = like_pattern_to_regex(pattern);
                 let re = BytesRegex::new(&re_pattern)
                     .expect("Unable to build regex from LIKE pattern: {}");
                 BooleanColumn::from_iterator(lhs.scalar_iter().map(|x| {
-                    let lhs_str = simdutf8::basic::from_utf8(x)
-                        .expect("Unable to convert lhs value to string: {}");
-                    let mut contain = false;
-                    for substring in sub_strings.clone() {
-                        contain = lhs_str.contains(substring);
-                    }
-                    if !contain {
-                        op(false)
+                    if !is_empty {
+                        let lhs_str = std::str::from_utf8(x)
+                            .expect("Unable to convert lhs value to string: {}");
+                        let contain = lhs_str.find(sub_strings[0]);
+                        if contain.is_none() {
+                            op(false)
+                        } else {
+                            op(re.is_match(x))
+                        }
                     } else {
                         op(re.is_match(x))
                     }