I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

split rhs pattern to rhs_vec: vec<string>

judge the lhs string contains the rhs_vec element.

if not contain, return false.

Fixes #8118