@TCeason 2022-10-14
use databend tpch 13 test on my local:

not optimize, first execute:
37 rows in set (0.73 sec)
Read 165000 rows, 10.47 MiB in 0.710 sec., 232.53 thousand rows/sec., 14.76 MiB/sec.


After optimize, first execute:
37 rows in set (0.67 sec)
Read 165000 rows, 10.47 MiB in 0.644 sec., 256.36 thousand rows/sec., 16.27 MiB/sec.


```sql
not optimize, first execute:

mysql> select     c_count,     count(*) as custdist from     (         select             c_custkey,             count(o_orderkey) as c_count         from             customer                 left outer join             orders             on c_custkey = o_custkey
       and o_comment not like '%pending%deposits%'         group by             c_custkey     )         c_orders group by     c_count order by     custdist desc,     c_count desc;
+---------+----------+
| c_count | custdist |
+---------+----------+
|       0 |     5000 |
|      10 |      676 |
|       9 |      651 |
|      11 |      618 |
|      12 |      554 |
|       8 |      548 |
|      13 |      514 |
|       7 |      487 |
|      19 |      485 |
|      18 |      461 |
|      14 |      454 |
|      20 |      444 |
|      16 |      442 |
|      17 |      438 |
|      15 |      430 |
|      21 |      396 |
|      22 |      378 |
|       6 |      355 |
|      23 |      322 |
|      24 |      262 |
|      25 |      188 |
|       5 |      184 |
|      26 |      162 |
|      27 |      138 |
|      28 |      103 |
|       4 |       92 |
|      29 |       59 |
|       3 |       49 |
|      30 |       29 |
|      31 |       26 |
|      32 |       19 |
|       2 |       12 |
|      33 |        8 |
|      34 |        7 |
|      35 |        5 |
|       1 |        3 |
|      36 |        1 |
+---------+----------+
37 rows in set (0.73 sec)
Read 165000 rows, 10.47 MiB in 0.710 sec., 232.53 thousand rows/sec., 14.76 MiB/sec.

After optimize, first execute:

mysql> select     c_count,     count(*) as custdist from     (         select             c_custkey,             count(o_orderkey) as c_count         from             customer                 left outer join             orders             on c_custkey = o_custkey
       and o_comment not like '%pending%deposits%'         group by             c_custkey     )         c_orders group by     c_count order by     custdist desc,     c_count desc;
+---------+----------+
| c_count | custdist |
+---------+----------+
|       0 |     5000 |
|      10 |      676 |
|       9 |      651 |
|      11 |      618 |
|      12 |      554 |
|       8 |      548 |
|      13 |      514 |
|       7 |      487 |
|      19 |      485 |
|      18 |      461 |
|      14 |      454 |
|      20 |      444 |
|      16 |      442 |
|      17 |      438 |
|      15 |      430 |
|      21 |      396 |
|      22 |      378 |
|       6 |      355 |
|      23 |      322 |
|      24 |      262 |
|      25 |      188 |
|       5 |      184 |
|      26 |      162 |
|      27 |      138 |
|      28 |      103 |
|       4 |       92 |
|      29 |       59 |
|       3 |       49 |
|      30 |       29 |
|      31 |       26 |
|      32 |       19 |
|       2 |       12 |
|      33 |        8 |
|      34 |        7 |
|      35 |        5 |
|       1 |        3 |
|      36 |        1 |
+---------+----------+
37 rows in set (0.67 sec)
Read 165000 rows, 10.47 MiB in 0.644 sec., 256.36 thousand rows/sec., 16.27 MiB/sec.

```