@xudong963 2022-10-14
I run the ticket with 10G tpch data on q13, the following is flamegraph

![f2](https://user-images.githubusercontent.com/41979257/195805398-2042d326-0110-4914-86f2-ed0b00c6e41b.svg)