@TCeason 2022-10-15
@xudong963  I load 10G data on my local pc:

run query

select * from orders where o_comment not like '%pending%deposits%';

in this pr the cost of stringsearchlike split to is_match_at and contains and from_utf8

![image](https://user-images.githubusercontent.com/33082201/195984914-98b7c2a2-29f7-4b7f-b910-1861f35d2832.png)

before this pr:

![image](https://user-images.githubusercontent.com/33082201/195984906-daad0187-737a-4281-91c9-2998656c279e.png)