@TCeason 2022-10-15
15.73s optimize to 11.05s

before optimize:

Database changed
mysql> select count(*) from orders where o_comment not like '%pending%deposits%';
+----------+
| count(*) |
+----------+
| 14838130 |
+----------+
1 row in set (15.73 sec)
Read 15000000 rows, 920.67 MiB in 15.685 sec., 956.31 thousand rows/sec., 58.70 MiB/sec.


after optimize:

Database changed
mysql> select count(*) from orders where o_comment not like '%pending%deposits%';
+----------+
| count(*) |
+----------+
| 14838130 |
+----------+
1 row in set (11.05 sec)
Read 15000000 rows, 920.67 MiB in 10.986 sec., 1.37 million rows/sec., 83.80 MiB/sec.