@TCeason 2022-10-18
> ```sql
> 'root'@mysqldb 11:22:43 [tpch]> select count(*) from orders where o_comment not like '%pending%deposits%';
> +----------+
> | count(*) |
> +----------+
> | 14838130 |
> +----------+
> 1 row in set (6.14 sec)
> Read 15000000 rows, 920.67 MiB in 6.091 sec., 2.46 million rows/sec., 151.15 MiB/sec.
> ```
> 
> Now the cost is almost on is_match. ![image](https://user-images.githubusercontent.com/33082201/196328472-4f229ae3-e900-40b3-b726-acda68e798db.png)
> 
> I try to use [bstr's find](https://docs.rs/bstr/0.1.0/bstr/struct.BStr.html#method.find) to instead of string::from_utf8 + string::find.
> 
> And the performance is not good. It cost about 9s.
> 
> So I also convert the x to string and then find the required_string.

If the modify is ok, I will modify the function-v2 like in this pr.